scripts/requirements.txt
duckdb>=1.0.0
requests>=2.31.0
pyyaml>=6.0
tabulate>=0.9.0
